A bunch of assorted changes
- Switched to stdint integer types

- Removed a compilation warning

- Added encoder build

+++ 26304_ANSI-C_source_code_v6_6_0/c-code/libbase/MP4Atoms.h
@@ -1240,7 +1240,7 @@ typedef struct MP4ItemPropertyContainerAtom {
 } MP4ItemPropertyContainerAtom, *MP4ItemPropertyContainerAtomPtr;
 
 typedef struct MP4ItemPropertyAssociationEntryPropertyIndex {
-	u8  essential;
+	u16  essential;
 	u16 property_index;
 } MP4ItemPropertyAssociationEntryPropertyIndex, *MP4ItemPropertyAssociationEntryPropertyIndexPtr;
